+++ r8s/r8s_service/adapter_client.py
@@ -587,6 +587,47 @@ class AdapterClient:
                                    method=HTTP_DELETE,
                                    payload=request)
 
+    def application_policies_get(self, application_id, group_id=None):
+        request = {
+            PARAM_APPLICATION_ID: application_id
+        }
+        if group_id:
+            request[PARAM_ID] = group_id
+
+        return self.__make_request(resource=API_APPLICATION_POLICIES,
+                                   method=HTTP_GET,
+                                   payload=request)
+
+    def application_policies_post(self, application_id, group_policy: dict):
+        request = {
+            PARAM_APPLICATION_ID: application_id,
+            **group_policy
+        }
+
+        return self.__make_request(resource=API_APPLICATION_POLICIES,
+                                   method=HTTP_POST,
+                                   payload=request)
+
+    def application_policies_patch(self, application_id, group_policy: dict):
+        request = {
+            PARAM_APPLICATION_ID: application_id,
+            **group_policy
+        }
+
+        return self.__make_request(resource=API_APPLICATION_POLICIES,
+                                   method=HTTP_PATCH,
+                                   payload=request)
+
+    def application_policies_delete(self, application_id, group_id=None):
+        request = {
+            PARAM_APPLICATION_ID: application_id,
+            PARAM_ID: group_id
+        }
+
+        return self.__make_request(resource=API_APPLICATION_POLICIES,
+                                   method=HTTP_DELETE,
+                                   payload=request)
+
     def parent_insights_resize(self, parent_id, instance_type):
         request = {
             PARAM_PARENT_ID: parent_id,
@@ -677,12 +718,10 @@ class AdapterClient:
                                    method=HTTP_PATCH,
                                    payload=request)
 
-    def shape_rule_delete(self, rule_id, parent_id=None):
+    def shape_rule_delete(self, rule_id):
         request = {
             PARAM_ID: rule_id
         }
-        if parent_id:
-            request[PARAM_PARENT_ID] = parent_id
 
         return self.__make_request(resource=API_SHAPE_RULES,
                                    method=HTTP_DELETE,