+++ r8s/CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.5.0] - 2024-01-24
+* Implement `r8s application policies` command group:
+  * `describe` - Describes available group policies from application
+  * `delete` - Deletes group policy from application
+  * `add_autoscaling` - Configures AUTO_SCALING group policy
+  * `update_autoscaling` - Updates AUTO_SCALING group policy
+* Remove unused `--parent_id` parameter from r8s parent shaperule delete
+* Fix invalid parent id in response of `r8s parent shaperule delete`
+
 ## [3.1.1] - 2024-01-18
 * fix issue with compatibility with `modular-api` for command `parent shape_rule`